Update rectangle direction comments and improve visualization
The comments specifying the direction of the rectangle in Rectangle.h have been updated to accurately define "1" as down and "-1" as up. Also, the visualization of corners of rectangles drawn on the console is improved in Render.cpp by using the ACS_ULCORNER, ACS_URCORNER, ACS_LLCORNER, and ACS_LRCORNER instead of '+'. Lastly, the code to calculate width and height of the rectangle has been uncommented in Rectangle.cpp along with minor improvements in rectangle's x position and speed calculation.