Refactor color handling in Ball class
Color vector is now passed as a pointer. The random color generator range has been fixed for appropriate color boundary limits. The constructor now receives a pointer to a vector which stores which colors are in use. Unused lines of code in the Render class have been removed for better efficiency and clarity.