Add bounce detection to Ball class
A new boolean variable 'justBounced' has been introduced to ensure that a bounce is only recorded once per collision with a rectangle. If the ball has just bounced, the execution of bounce-related logic is prevented until the ball is no longer in the rectangle. This prevents over-counting of bounces and erratic bounce behaviour.