Part 3 works, but with busy waiting
The Ball class was modified to support bouncing queue logic. A waiting queue system has been implemented where balls that have bounced from the rectangle are stored. This functionality was added to both the Ball and Render classes and is expected to enhance the handling of ball and rectangle interactions.