Implement condition variables for ball queue
The ball queue system has been updated to now use a condition variable, improving the efficiency of the communication between threads. The 'notifyAllBalls' function has been introduced, which is uded to activate the condition variable when ending the program. The ball waiting process in 'runBall' has been reworked to use 'queueCV.wait', reducing the reliance on a sleep-then-check loop and organically improving the system.