Add maximum ball limit and adjust spawn rate
Modified the `Render` module to limit the number of active balls to 20 through the addition of a `MAX_BALLS` constant. Also adjusted the random distribution parameters in the `spawnBall` function to increase the ball spawn rate. The balls' storage was changed from queue to vector type for ease of size checking.