Ukończenie projektu (#2)
* simpler version

* Implement collision checks in ball movement

Improved the collision detection logic in Ball::runBall() by correctly identifying intersections between different movement directions and rectangle boundaries. Added new method calls to access attributes of Rectangle objects. This update also introduces changes to the sleep duration control in both Ball and Rectangle classes, implementing more accurate and robust mechanics.

* Refactor collision detection and modify ball speed calculation

The existing verbose comments on collision detection were removed and the collision detection code has been refactored for simplicity and readability. Removed the 'justBounced' boolean flag and introduced 'insideCounter' to record consecutive collision occurrences.

* Add 'etap2.txt' with stage 2 elements

* Add bounce detection to Ball class

A new boolean variable 'justBounced' has been introduced to ensure that a bounce is only recorded once per collision with a rectangle. If the ball has just bounced, the execution of bounce-related logic is prevented until the ball is no longer in the rectangle. This prevents over-counting of bounces and erratic bounce behaviour.

* Refactor rectangle edge calculation and speed generation

Edges calculation for the Rectangle class has been commented out as it is not used in this version.

* Add mutex library to Rectangle.h

* Update rules for ball bouncing

Extended the description of the bouncing rules in etap2.txt. Added detailed information about the change in ball direction after bouncing, speed adjustment mechanism, and the scenario when a ball gets trapped between a rectangle and a wall.

* Add 'bouncedFromRectangle' condition and visualization

A new 'bouncedFromRectangle' condition has been added to the Ball class to accurately track when it bounces off a rectangle. This condition is then utilized in the 'Render' class to visually indicate such instances, changing the drawn character for a ball that has bounced off a rectangle.

* Part 3 works, but with busy waiting

The Ball class was modified to support bouncing queue logic. A waiting queue system has been implemented where balls that have bounced from the rectangle are stored. This functionality was added to both the Ball and Render classes and is expected to enhance the handling of ball and rectangle interactions.

* Add thread-safe operations to Ball class

Mutex protection was added to various queue operations in the Ball class to ensure thread safety. Specifically, several queue manipulations within the Ball class methods are now done within a locked context using a mutex. This prevents potential race conditions and ensures consistent state of the Ball instances.

* Refactor Ball class to improve thread safety

The Ball class's destructor and runBall method has been updated to check if the waitingBalls queue is empty before accessing its front. Furthermore, an additional condition was added in runBall to ensure a ball won't wait in the queue if it's within the range of the rectangle. This prevents potential crashes due to attempts to access non-existing elements, enhancing the thread safety and stability of the application.

* Change ball waiting queue to vector

The code has been updated to change the data structure used for storing balls that are waiting after bouncing from a rectangle from queue to vector.

* Implement condition variables for ball queue

The ball queue system has been updated to now use a condition variable, improving the efficiency of the communication between threads. The 'notifyAllBalls' function has been introduced, which is uded to activate the condition variable when ending the program. The ball waiting process in 'runBall' has been reworked to use 'queueCV.wait', reducing the reliance on a sleep-then-check loop and organically improving the system.

* Add separate mutex for condition variable in Ball class

A separate mutex has been introduced for the condition variable queueCV to avoid potential issues with synchronization in the Ball class. Previously, the mutex for the Ball queue was being used for condition variable wait, which can now be prevented.

* Add maximum ball limit and adjust spawn rate

Modified the `Render` module to limit the number of active balls to 20 through the addition of a `MAX_BALLS` constant. Also adjusted the random distribution parameters in the `spawnBall` function to increase the ball spawn rate. The balls' storage was changed from queue to vector type for ease of size checking.

* Change condition variable notification from one to all.

* Remove commented out code from several files

This commit removes unused, commented out code from Ball.cpp, Rectangle.h, Rectangle.cpp and Render.cpp.

* Add condition_variable library to Ball.h

The condition_variable library has been included in Ball.h.